+++ src/main/java/com/bookspot/batch/infra/opensearch/OpenSearch504Exception.java
@@ -1,4 +1,4 @@
-package com.bookspot.batch.step.service.opensearch;
+package com.bookspot.batch.infra.opensearch;
 
 public class OpenSearch504Exception extends RuntimeException {
     public OpenSearch504Exception(Throwable cause) {
