+++ src/main/java/com/bookspot/batch/step/BookOpenSearchSyncStepConfig.java
@@ -7,8 +7,8 @@ import com.bookspot.batch.step.listener.StepLoggingListener;
 import com.bookspot.batch.step.listener.alert.AlertStepListener;
 import com.bookspot.batch.step.reader.BookWithLibraryIdReader;
 import com.bookspot.batch.step.service.*;
-import com.bookspot.batch.step.service.opensearch.OpenSearch504Exception;
-import com.bookspot.batch.step.service.opensearch.OpenSearchRepository;
+import com.bookspot.batch.infra.opensearch.OpenSearch504Exception;
+import com.bookspot.batch.infra.opensearch.OpenSearchRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.StepScope;
